Added the ability to mutate vectors ala 'vector-set!'

+++ c-src/constants.h
@@ -1,9 +1,9 @@
 #ifndef LC_CONSTS
 #define LC_CONSTS
 
-const int TRUE_VALUE;
-const int FALSE_VALUE;
-const int NIL_VALUE;
+const unsigned long TRUE_VALUE;
+const unsigned long FALSE_VALUE;
+const unsigned long NIL_VALUE;
 
 const int FIXNUM_TAG;
 const int FIXNUM_TAG_LEN;
